+++ DifferentiationInterfaceTest/src/utils.jl
@@ -1,4 +1,4 @@
-mysimilar(x::Number) = one(x)
+mysimilar(x::Number) = oneunit(x)
 mysimilar(x::AbstractArray) = similar(x)
 mysimilar(x::Union{Tuple,NamedTuple}) = map(mysimilar, x)